handle parsing errors
