+++ src/parser/tree_get_command.c
@@ -1,9 +1,5 @@
 #include <minishell.h>
 
-void	consume(t_token_lst **token_lst)
-{
-	(*token_lst) = (*token_lst)->next;
-}
 
 t_tree	*tree_get_io_redirect_list(t_token_lst	**token_lst)
 {
@@ -11,7 +7,7 @@ t_tree	*tree_get_io_redirect_list(t_token_lst	**token_lst)
 	t_tree	*infiles;
 	t_tree	*outfiles;
 
-		printf("%s\n", (char *)(*token_lst)->token->lexeme);
+	printf("%s\n", (char *)(*token_lst)->token->lexeme);
 	io_redirect_list = tree_create_new(6, NULL);
 	infiles = NULL;
 	outfiles = NULL;
@@ -39,26 +35,29 @@ t_tree	*tree_get_io_redirect_list(t_token_lst	**token_lst)
 
 t_tree	*parse_subshell(t_token_lst	**token_lst)
 {
-	t_tree	*simple_command;
+	t_tree	*compound_command;
 	t_tree	*io_redirect_list;
 	t_tree	*subshell;
 
-	consume(token_lst);
-	simple_command = tree_get_command(token_lst);
-	//if (!(*token_lst) || (!simple_command->next->next && !simple_command->next->sibling->next))
-	//{
-	//printf("error\n");
-	//exit (2);
-	//}
-	if ((*token_lst)->token->type == R_PAREN)
-		consume(token_lst);
-	subshell = tree_create_new(7, NULL);
+	subshell = NULL;
 	if ((*token_lst))
 	{
-		io_redirect_list = tree_get_io_redirect_list(token_lst);
-		tree_add_sibling_back(&subshell, io_redirect_list);
+		consume(token_lst);
+		compound_command = tree_get_compound_command(token_lst);
+		if (compound_command == NULL)
+			printf("error expected '('");
+		if ((*token_lst) && (*token_lst)->token->type == R_PAREN)
+			consume(token_lst);
+		else
+			printf("error expected '('");
+		subshell = tree_create_new(7, NULL);
+		if ((*token_lst))
+		{
+			io_redirect_list = tree_get_io_redirect_list(token_lst);
+			tree_add_sibling_back(&subshell, io_redirect_list);
+		}
+		tree_add_back(&subshell, compound_command);
 	}
-	tree_add_back(&subshell, simple_command);
 	return (subshell);
 }
 
@@ -68,19 +67,25 @@ t_tree	*tree_get_command(t_token_lst **token_lst)
 	t_tree	*subshell;
 	t_tree	*simple_command;
 
-	command = tree_create_new(8, NULL);
-	simple_command = NULL;
-	//for subshell: consume '(' and call tree_getcommand_list and then consume ')'
-	if ((*token_lst)->token->type == L_PAREN)
+	command = NULL;
+	if ((*token_lst))
 	{
-		subshell = parse_subshell(token_lst);
-		tree_add_back(&command, subshell);
-	}
-	//for simple command: 
-	else
-	{
-		simple_command = parse_simple_command(token_lst);
-		tree_add_back(&command, simple_command);
+		command = tree_create_new(8, NULL);
+		simple_command = NULL;
+		//for subshell: consume '(' and call tree_getcommand_list and then consume ')'
+		if ((*token_lst)->token->type == L_PAREN)
+		{
+			subshell = parse_subshell(token_lst);
+			tree_add_back(&command, subshell);
+		}
+		//for simple command: 
+		else
+		{
+			simple_command = parse_simple_command(token_lst);
+			if (simple_command == NULL)
+				return (NULL);
+			tree_add_back(&command, simple_command);
+		}
 	}
 	return (command);
 }
