+++ src/parser/tree_get_pipeline.c
@@ -10,16 +10,18 @@ t_tree	*tree_get_pipeline(t_token_lst **token_lst)
 	commands = NULL;
 	if ((*token_lst))
 	{
-		pipeline = tree_create_new(9, NULL);
 		while ((*token_lst))
 		{
 			command = tree_get_command(token_lst);
+			if (command == NULL)
+				return (NULL);
 			tree_add_sibling_back(&commands, command);
-			if ((*token_lst) && (*token_lst)->token->type == PIPE)
+			if ((*token_lst) && (*token_lst)->token->type == PIPE && (*token_lst)->next)
 				consume(token_lst);
 			else
 				break ;
 		}
+		pipeline = tree_create_new(9, NULL);
 		tree_add_back(&pipeline, commands);
 	}
 	return (pipeline);
