+++ src/parser/tree_utils.c
@@ -54,24 +54,3 @@ void	tree_add_sibling_back(t_tree **head, t_tree *new)
 		last_sibling->sibling = new;
 	}
 }
-
-
-//int	main(void)
-//{
-	//t_tree *head;
-	//head = NULL;
-	////t_tree *hey = tree_create_new(4, "hey hey");
-	//tree_add_back(&head, tree_create_new(3, "number 1"));
-	//tree_add_back(&head, tree_create_new(4, "number 2"));
-	//tree_add_back(&head, tree_create_new(4, "number 3"));
-	//tree_add_sibling_back(&head, tree_create_new(4, "I'm one of siblings"));
-	//printf("%s\n", head->sibling->data);
-	//printf("%s\n", head->data);
-	//printf("%s\n", head->next->data);
-	//printf("%s\n", (tree_get_last(head)->data));
-	//free(head->sibling);
-	//free(head->next->next);
-	//free(head->next);
-	//free(head);
-	//return (0);
-//}
